requirements.txt
asgiref==3.7.2
attrs==24.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
CacheControl==0.14.0
cachetools==5.5.2
certifi==2025.10.5
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.4.4
constantly==15.1.0
cryptography==45.0.7
daphne==4.0.0
Django==3.2.25
django-ckeditor==5.9.0
django-environ==0.11.2
django-js-asset==1.2.2
django-model-utils==4.0.0
djangorestframework==3.15.1
firebase-admin==6.8.0
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.62.3
grpcio-status==1.62.3
httplib2==0.31.0
hyperlink==21.0.0
idna==3.10
importlib-metadata==6.7.0
incremental==22.10.0
msgpack==1.0.5
Pillow==7.0.0
proto-plus==1.26.1
protobuf==4.24.4
psycopg2==2.8.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==25.3.0
pyparsing==3.1.4
pytz==2019.3
requests==2.31.0
rsa==4.9.1
service-identity==21.1.0
six==1.17.0
sqlparse==0.3.1
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.4.0
Unipath==1.1
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.15.0
zope.interface==6.4.post2
